+++ halley-retrofit/src/main/kotlin/com/infinum/halley/retrofit/converters/options/HalleyQueryArgumentsFactory.kt
@@ -23,9 +23,9 @@ internal class HalleyQueryArgumentsFactory :
                     annotationParameters(it.arguments) + cacheParameters(it.key)
                 )
             } else {
-                null
+                HalleyOptions.query()
             }
-        }
+        } ?: HalleyOptions.query()
 
     override fun annotationParameters(parameters: Array<HalQueryArgument>): HalleyKeyedMap =
         parameters