+++ CHANGELOG.md
@@ -1,9 +1,19 @@
 Changelog
 =========
 
+## Version 0.0.2
+
+_2023-01-12_
+
+* Update dependencies
+* Fix serialization deprecation.
+* Implement sanitization of templated links without value for placeholders.
+* Return imperative arguments when there are no annotated arguments present.
+* Implement common template arguments.
+
 ## Version 0.0.1
 
 _2022-08-16_
 
-Initial public release.
+* Initial public release.